chore: Extract keyUtils to be used in other key-based collection hooks
